+++ azext_iot/tests/iothub/core/test_iot_messaging_int.py
@@ -130,7 +130,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         self.kwargs["messaging_unicodable_data"] = read_file_content(messaging_unicodable_data_path)
         self.kwargs["messaging_non_unicodable_data"] = "{{non-decodable content}}"
         self.cmd(
-            """iot device c2d-message send -d {} -n {} -g {} --data '{}' --cid {} --mid {} --ct {} --expiry {}
+            """iot device c2d-message send -d {} -n {} -g {} --fp '{}' --cid {} --mid {} --ct {} --expiry {}
             --ce {} --props {} -y""".format(
                 device_ids[0],
                 self.entity_name,
@@ -183,7 +183,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # Send C2D message with binary payload in file
         self.cmd(
-            """iot device c2d-message send -d {} -n {} -g {} --data '{}' --cid {} --mid {} --ct {} --expiry {}
+            """iot device c2d-message send -d {} -n {} -g {} --fp '{}' --cid {} --mid {} --ct {} --expiry {}
             --ce {} --props {} -y""".format(
                 device_ids[0],
                 self.entity_name,
@@ -235,7 +235,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # Send C2D message with non-unicode decodable binary payload in file
         self.cmd(
-            """iot device c2d-message send -d {} -n {} -g {} --data '{}' --cid {} --mid {} --ct {} --expiry {}
+            """iot device c2d-message send -d {} -n {} -g {} --fp '{}' --cid {} --mid {} --ct {} --expiry {}
             --ce {} --props {} -y""".format(
                 device_ids[0],
                 self.entity_name,
@@ -287,7 +287,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # Send C2D message with non-unicode content encoding
         self.cmd(
-            """iot device c2d-message send -d {} -n {} -g {} --data '{}' --cid {} --mid {} --ct {} --expiry {}
+            """iot device c2d-message send -d {} -n {} -g {} --fp '{}' --cid {} --mid {} --ct {} --expiry {}
             --ce {} --props {} -y""".format(
                 device_ids[0],
                 self.entity_name,
@@ -335,6 +335,24 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
                 device_ids[0], self.entity_name, self.entity_rg, etag
             ),
             checks=self.is_empty(),
+        )
+
+        # Error - Send C2D message with non existed file path
+        self.cmd(
+            """iot device c2d-message send -d {} -n {} -g {} --fp '{}' --cid {} --mid {} --ct {} --expiry {}
+            --ce {} --props {} -y""".format(
+                device_ids[0],
+                self.entity_name,
+                self.entity_rg,
+                '123',
+                test_cid,
+                test_mid,
+                'application/json',
+                test_et,
+                test_ce,
+                test_props,
+            ),
+            expect_failure=True,
         )
 
         # Send C2D message via --login + application/json content ype
@@ -550,7 +568,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         device_events.append((device_id, send_d2c_msg))
 
         self.cmd(
-            "iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/json;$.ce=utf-8' --key {}".format(
+            "iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/json;$.ce=utf-8' --key {}".format(
                 device_id, self.entity_name, self.entity_rg, messaging_data_path,
                 keys["primaryKey"]
             )
@@ -581,7 +599,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # Send-d2c-message with secondary key and binary data
         self.cmd(
-            "iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/octet-stream' --key {}".format(
+            "iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream' --key {}".format(
                 device_id, self.entity_name, self.entity_rg, messaging_unicodable_data_path,
                 keys["secondaryKey"]
             )
@@ -589,7 +607,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         device_events.append((device_id, self.kwargs["messaging_unicodable_data"]))
 
         self.cmd(
-            "iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/octet-stream' --key {}".format(
+            "iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream' --key {}".format(
                 device_id, self.entity_name, self.entity_rg, messaging_non_unicodable_data_path,
                 keys["secondaryKey"]
             )
@@ -613,6 +631,15 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         self.cmd(
             "iot device send-d2c-message -d {} -n {} -g {} --da '{}' --key {} --model-id {} ".format(
                 device_id, self.entity_name, self.entity_rg, send_d2c_msg, keys["primaryKey"], "dtmi:com:example;"
+            ),
+            expect_failure=True
+        )
+
+        # Error - send-d2c-message with non existed file path.
+        self.cmd(
+            "iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream' --key {}".format(
+                device_id, self.entity_name, self.entity_rg, '123',
+                keys["secondaryKey"]
             ),
             expect_failure=True
         )
@@ -672,7 +699,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         device_events.append((device_ids[0], send_d2c_msg))
 
         self.cmd(
-            """iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/json;$.ce=utf-8'
+            """iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/json;$.ce=utf-8'
             --cp {} --kp {}""".format(
                 device_ids[0], self.entity_name, self.entity_rg, messaging_data_path,
                 f"{device_ids[0]}-cert.pem", f"{device_ids[0]}-key.pem"
@@ -748,7 +775,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         device_events.append((device_ids[1], send_d2c_msg))
 
         self.cmd(
-            """iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/octet-stream'
+            """iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream'
             --cp {} --kp {} --pass {}""".format(
                 device_ids[1], self.entity_name, self.entity_rg, messaging_unicodable_data_path,
                 f"{device_ids[1]}-cert.pem", f"{device_ids[1]}-key.pem", fake_pass
@@ -757,13 +784,23 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         device_events.append((device_ids[1], self.kwargs["messaging_unicodable_data"]))
 
         self.cmd(
-            """iot device send-d2c-message -d {} -n {} -g {} --da '{}' -p '$.ct=application/octet-stream'
+            """iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream'
             --cp {} --kp {} --pass {}""".format(
                 device_ids[1], self.entity_name, self.entity_rg, messaging_non_unicodable_data_path,
                 f"{device_ids[1]}-cert.pem", f"{device_ids[1]}-key.pem", fake_pass
             )
         )
         device_events.append((device_ids[1], self.kwargs["messaging_non_unicodable_data"]))
+
+        # Error - send-d2c-message with non existed file path.
+        self.cmd(
+            """iot device send-d2c-message -d {} -n {} -g {} --fp '{}' -p '$.ct=application/octet-stream'
+            --cp {} --kp {} --pass {}""".format(
+                device_ids[1], self.entity_name, self.entity_rg, '123',
+                f"{device_ids[1]}-cert.pem", f"{device_ids[1]}-key.pem", fake_pass
+            ),
+            expect_failure=True
+        )
 
         self._monitor_checker(enqueued_time=enqueued_time, device_events=device_events)
 
@@ -1276,7 +1313,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
             cmd=client,
             device_id=device_ids[i],
             hub_name=self.entity_name,
-            data=messaging_non_unicodable_data_path,
+            file_path=messaging_non_unicodable_data_path,
             properties="$.ct=application/octet-stream",
             msg_count=1,
             resource_group_name=self.entity_rg,
@@ -1284,7 +1321,7 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # Monitor events for a single device contains non-unicode decodable message body
         self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 8 -y -p all".format(
+            "iot hub monitor-events -n {} -g {} --cg {} --et {} -y -p all".format(
                 self.entity_name, self.entity_rg, LIVE_CONSUMER_GROUPS[1], enqueued_time
             ),
             ['{{non-unicode decodable payload}}'],